requirements.txt
albumentations==1.4.4
batchgenerators==0.25
easydict==1.11
MedPy==0.5.1
monai==1.3.0
nibabel==5.3.2
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.1.4
prefetch_generator==1.0.3
pycocotools==2.0.7
SimpleITK==2.4.1
torch==2.0.0
torchinfo
torchshow
torchvision
tqdm
Pillow==9.3.0
numpy==1.26.3
openpyxl==3.1.2
